refactor: rename upgrade v5 to v6
