+++ app/app.go
@@ -42,7 +42,7 @@ import (
 	v2 "github.com/sge-network/sge/app/upgrades/v2"
 	v3 "github.com/sge-network/sge/app/upgrades/v3"
 	v4 "github.com/sge-network/sge/app/upgrades/v4"
-	v5 "github.com/sge-network/sge/app/upgrades/v5"
+	v6 "github.com/sge-network/sge/app/upgrades/v6"
 	abci "github.com/tendermint/tendermint/abci/types"
 	tmjson "github.com/tendermint/tendermint/libs/json"
 	"github.com/tendermint/tendermint/libs/log"
@@ -76,7 +76,7 @@ var (
 		v2.Upgrade,
 		v3.Upgrade,
 		v4.Upgrade,
-		v5.Upgrade,
+		v6.Upgrade,
 	}
 )
