+++ app/upgrades/v6/consts.go
@@ -1,4 +1,4 @@
-package v5
+package v6
 
 import (
 	store "github.com/cosmos/cosmos-sdk/store/types"
