+++ app/upgrades/v6/upgrades.go
@@ -1,4 +1,4 @@
-package v5
+package v6
 
 import (
 	"time"
